fix(readiness): P2 underbuilt — do.ts anyOf parity, EUR-Lex challenge escalation, ToS gate in the fetch layer
Three items from the P1 handoff worklist, none needing human decisions:

- /v1/do input validation now has anyOf/oneOf group parity with the x402
  surface (the PR #171 incident class): the classic-required check kept its
  established hints/analytics untouched; the group rule is added after it via
  the same pure validateX402Input both x402 handlers use. Failures log
  failure_type=input_group_unsatisfied and return the schema.
- eu-regulation-search fixed (was 100% failing in sweeps): EUR-Lex now serves
  a JS anti-bot interstitial that squeaked past tier-1's substantial-content
  heuristic (2036 bytes > 2000; embedded CSS counted toward body text), so the
  plain fetch "succeeded" with the challenge shell and the browser tier never
  ran. web-provider now detects challenge interstitials (narrow markers) and
  escalates them to the rendered browser, which satisfies the JS requirement
  legitimately; if even the rendered browser is challenged, it fails honestly
  — Strale does not attempt to defeat bot walls. Also: style content no
  longer counts toward the body-text heuristic; the executor texts the
  SearchResult region instead of the nav-heavy page head, and passes
  networkidle2 + wider budgets (fetchRenderedHtml now forwards options).
  Verified live: 10 regulations, first hit Regulation (EU) 2024/1689.
- ToS blocklist moved to src/lib (re-export shim keeps all 9 executor imports
  stable — the company-name-match PR #161 precedent) and enforced at BOTH
  shared entry points: safeFetch (including per-redirect-hop re-checks, so
  bit.ly → linkedin.com can't slip through) and fetchPage (tiers 2/3 never
  touch safeFetch). A new URL capability can no longer ship ungated by
  omission. Per-executor calls stay as defense in depth.

tsc clean; 1071 tests pass (13 new).
