+++ apps/api/src/lib/safe-fetch.ts
@@ -27,6 +27,7 @@ import { Agent as HttpsAgent } from "node:https";
 import { Agent as HttpAgent } from "node:http";
 import { lookup as dnsLookup, type LookupOptions } from "node:dns";
 import { isBlockedIp, validateUrl } from "./url-validator.js";
+import { assertTargetAllowed } from "./tos-blocklist.js";
 import { logWarn } from "./log.js";
 
 // ─── Custom lookup that re-applies isBlockedIp ────────────────────────────────
@@ -141,6 +142,13 @@ export async function safeFetch(
 ): Promise<Response> {
   const { maxRedirects = DEFAULT_MAX_REDIRECTS, timeoutMs, ...init } = opts;
 
+  // Per-source ToS policy is enforced at the fetch layer, not only in the
+  // capabilities that remember to call it (P2, 2026-08-12 — three separate
+  // capabilities had shipped ungated). Pure string check, throws with
+  // TOS_REFUSAL_MARKER so breakers classify it as caller error and DEC-14
+  // leaves the caller unbilled. Redirect hops re-check below.
+  assertTargetAllowed(String(url));
+
   // Cert-audit C6: ensure every safeFetch caller is timeout-bounded so a
   // stuck upstream can't pin a connection or pool slot indefinitely.
   // Caller's own signal wins if provided; otherwise apply DEFAULT_TIMEOUT_MS.
@@ -210,5 +218,8 @@ export async function followRedirects(
     }
 
     current = new URL(location, current).toString();
+    // A redirect into a ToS-prohibited host is the same policy violation as
+    // fetching it directly (bit.ly → linkedin.com must not slip through).
+    assertTargetAllowed(current);
   }
 }
